fix(backend): using int64 instead of unit64 for time
Im pretty sure linus was a 1y old back then

+++ backend/internal/git/get_blame.go
@@ -40,7 +40,7 @@ func GetBlame(repoId uint64, commitHash string, path string) ([]BlameLine, error
 				Email: line.Author,
 			},
 			Text:      line.Text,
-			Timestamp: uint64(line.Date.Unix()),
+			Timestamp: line.Date.Unix(),
 			Hash:      line.Hash.String(),
 		})
 	}
